+++ lib/runtime-core/src/cache.rs
@@ -1,4 +1,5 @@
 use crate::{
+    backend::Backend,
     module::{Module, ModuleInfo},
     sys::Memory,
 };
@@ -36,17 +37,31 @@ impl From<io::Error> for Error {
 pub struct WasmHash([u8; 32], [u8; 32]);
 
 impl WasmHash {
-    /// Hash a wasm module.
+    /// Hash a wasm module for the default compiler backend.
+    ///
+    /// See also: `WasmHash::generate_for_backend`.
     ///
     /// # Note:
     /// This does no verification that the supplied data
     /// is, in fact, a wasm module.
     pub fn generate(wasm: &[u8]) -> Self {
+        WasmHash::generate_for_backend(wasm, Backend::default())
+    }
+
+    /// Hash a wasm module for a specific compiler backend.
+    /// This allows multiple cache entries containing the same compiled
+    /// module with different compiler backends.
+    ///
+    /// # Note:
+    /// This function also does no verification that the supplied data
+    /// is a wasm module
+    pub fn generate_for_backend(wasm: &[u8], backend: Backend) -> Self {
         let mut first_part = [0u8; 32];
         let mut second_part = [0u8; 32];
 
         let mut state = blake2bp::State::new();
         state.update(wasm);
+        state.update(backend.to_string().as_bytes());
 
         let hasher = state.finalize();
         let generic_array = hasher.as_bytes();